Spatial filtering techniques applied.
Filter term in “Digital image processing” is referred to the subimage. Spatial filtering term is the filtering operations that are performed directly on the pixels of an image. The technique is used for changing the intensities of a pixel according to the intensities of the neighboring pixels. Using spatial filtering, the image is transformed  based on a kernel H which has certain height and width.  In this assignment, we are asked to use some MATLAB’s functions related to filtering, smoothing, sharpening etc. For part 1, I implemented correlation operation  which will return to same output as MATLAB’s built-in imfilter function. Secondly, on part 2 I used the function recently implemented for the correlation of an image and a template. After that, I compared averaging filter and Gaussian filter for the third part. Then, to complete part 4, I applied median fitering to a corrupted image. For the next part (part 5), I implemented unsharp masking. Finally, for the bonus part I used bilateral filter function. The detailed explanation of the steps are in the report.